requirements.txt
# ─── CosmoERP — Requirements ──────────────────────────────────────────────────

# Framework Django
Django>=4.2,<5.0
djangorestframework>=3.14

# Filtering
django-filter>=23.0

# Database
psycopg2-binary>=2.9.9

# Configuration
python-decouple>=3.8

# Images
Pillow>=10.0

# Static files (WhiteNoise pour Railway/production)
whitenoise>=6.6.0

# CORS
django-cors-headers>=4.3.0

# Supabase Python SDK (optionnel - pour fonctionnalités avancées)
supabase>=2.4.0

# Caching & Performance
django-redis>=5.4.0

# Monitoring
sentry-sdk>=1.39.0

# Export Excel/PDF
openpyxl>=3.1.2
reportlab>=4.1.0

# JWT Authentication
djangorestframework-simplejwt>=5.3.0

# AI Assistant
anthropic>=0.25.0

# WSGI server (pour Railway/production)
gunicorn>=21.2.0

# WebSocket (Django Channels - notifications temps réel)
channels>=4.0.0
channels-redis>=4.1.0
